os/board/rtl8730e: Revise data retrieve flow from external flash
- Revise data retrieve flow from external flash
- Revise the condition checking for data binary
