+++ os/board/rtl8730e/src/component/bluetooth/driver/platform/amebad2/hci/hci_platform.c
@@ -12,6 +12,18 @@
 #include "hci_dbg.h"
 #include "dlist.h"
 
+#ifdef CONFIG_SECOND_FLASH_PARTITION
+#include <tinyara/spi/spi.h>
+#include "spi_api.h"
+
+/* Corresponds to LD external flash starting address */
+#define EXT_FLASH_BASE_ADDRESS   			0x09000000
+#define HCI_PATCH_ADDR_OFFSET				0x0
+#define EXTERNAL_FLASH_READ_CMD_SIZE		5
+/* For getting total binary size and addr */
+#define EXTERNAL_FLASH_BINARY_HEADER_SIZE	16
+#endif
+
 #define HCI_PHY_EFUSE_OFFSET       0x740
 #define HCI_PHY_EFUSE_LEN          0x70
 #define HCI_LGC_EFUSE_OFFSET       0x1B0
@@ -924,8 +936,58 @@ static uint32_t hci_platform_parse_patch(uint8_t *p_buf, HCI_PATCH_NODE *p_patch
 }
 
 #ifdef CONFIG_SECOND_FLASH_PARTITION
-extern uint8_t *rtlbt_fw_ptr;
-extern uint32_t rtlbt_fw_size;
+static void external_flash_read_stream(struct spi_dev_s *spi, char *rx_buffer, uint32_t length, uint32_t offset)
+{
+    char cmd[EXTERNAL_FLASH_READ_CMD_SIZE] = {0};
+    /* transmit read cmd */
+    cmd[0] = 0x13;
+    cmd[1] = (offset >> 24) & 0xFF;
+    cmd[2] = (offset >> 16) & 0xFF;
+    cmd[3] = (offset >> 8) & 0xFF;
+    cmd[4] = offset & 0xFF;
+
+    SPI_EXCHANGE(spi, (char *)cmd, (char *)rx_buffer, length + EXTERNAL_FLASH_READ_CMD_SIZE);
+    memmove(rx_buffer, rx_buffer + EXTERNAL_FLASH_READ_CMD_SIZE, length);
+}
+
+static uint8_t* hci_platform_get_btfw_patch(uint32_t *rtlbt_fw_size)
+{
+    /* SPI Initialize */		
+    struct spi_dev_s *spi;
+    uint8_t *rtlbt_fw_ptr = NULL;
+    spi = up_spiinitialize(1);
+    SPI_LOCK(spi, true);
+    SPI_SETBITS(spi, 8);
+    SPI_SETFREQUENCY(spi, 1000000);
+    uint8_t *rtlbt_fw_hdr_ptr = NULL;
+    uint32_t rtlbt_fw_addr = 0;
+
+    /* SPI Read size*/
+    rtlbt_fw_hdr_ptr = osif_mem_alloc(RAM_TYPE_DATA_ON, EXTERNAL_FLASH_BINARY_HEADER_SIZE);
+    SPI_SELECT(spi, 1, true);
+    external_flash_read_stream(spi, (char *)rtlbt_fw_hdr_ptr, EXTERNAL_FLASH_BINARY_HEADER_SIZE, HCI_PATCH_ADDR_OFFSET);
+    SPI_SELECT(spi, 1, false);
+    rtlbt_fw_addr = (rtlbt_fw_hdr_ptr[12]) | (rtlbt_fw_hdr_ptr[13] << 8) | (rtlbt_fw_hdr_ptr[14] << 16) | (rtlbt_fw_hdr_ptr[15] << 24);
+    *rtlbt_fw_size = (rtlbt_fw_hdr_ptr[8]) | (rtlbt_fw_hdr_ptr[9] << 8) | (rtlbt_fw_hdr_ptr[10] << 16) | (rtlbt_fw_hdr_ptr[11] << 24);
+    /* Below condition is minimal requirement to meet */
+	/* Address must be starting of the external flash, size shouldn't be larger than 128KB */
+    if(!((*rtlbt_fw_size > 0 && *rtlbt_fw_size < 0x20000) && (rtlbt_fw_addr == EXT_FLASH_BASE_ADDRESS))) {
+        dbg("Please check whether RTK data binary is flashed correctly!");
+        ASSERT(0);
+    }
+    if (rtlbt_fw_hdr_ptr) {
+    	osif_mem_free(rtlbt_fw_hdr_ptr);
+    }
+    /* SPI Read Data back */
+    rtlbt_fw_ptr = osif_mem_alloc(RAM_TYPE_DATA_ON, *rtlbt_fw_size);
+    SPI_SELECT(spi, 1, true);
+    /* The binary header size is 32bytes, so we need to add back the offset */
+    external_flash_read_stream(spi, (char *)rtlbt_fw_ptr, *rtlbt_fw_size, rtlbt_fw_addr - (EXT_FLASH_BASE_ADDRESS + HCI_PATCH_ADDR_OFFSET) + 32);
+    SPI_SELECT(spi, 1, false);
+    SPI_LOCK(spi, false);
+
+    return rtlbt_fw_ptr;
+}
 #endif
 
 static uint8_t hci_platform_get_patch_info(void)
@@ -938,6 +1000,12 @@ static uint8_t hci_platform_get_patch_info(void)
 	uint8_t project_id;
 	uint32_t version_date, version_time;
 	uint32_t fw_len;
+	uint8_t ret = HCI_SUCCESS;
+
+#ifdef CONFIG_SECOND_FLASH_PARTITION
+	uint32_t rtlbt_fw_size;
+	uint8_t *rtlbt_fw_ptr = NULL;
+#endif
 
 	if (CHECK_CFG_SW(CFG_SW_USE_FLASH_PATCH)) {
 #if defined(CONFIG_BT_EXCLUDE_MP_FUNCTION) && CONFIG_BT_EXCLUDE_MP_FUNCTION
@@ -946,6 +1014,7 @@ static uint8_t hci_platform_get_patch_info(void)
 		patch_info->patch_len = rtlbt_fw_len;
 #else
 		/* Assign back to BT structure */
+		rtlbt_fw_ptr = hci_platform_get_btfw_patch(&rtlbt_fw_size);
 		patch_info->patch_buf = (uint8_t *)(void *)rtlbt_fw_ptr;
 		patch_info->patch_len = rtlbt_fw_size;
 #endif
@@ -968,7 +1037,8 @@ static uint8_t hci_platform_get_patch_info(void)
 		if ((!memcmp(patch_info->patch_buf, patch_sig_v1, sizeof(patch_sig_v1))) && \
 			(!memcmp(patch_info->patch_buf + patch_info->patch_len - sizeof(ext_section_sig), ext_section_sig, sizeof(ext_section_sig)))) {
 			HCI_WARN("Signature check success: Merge patch v1 not support");
-			return HCI_IGNORE;
+			ret = HCI_IGNORE;
+			goto EXIT;
 		} else if ((!memcmp(patch_info->patch_buf, patch_sig_v2, sizeof(patch_sig_v2))) && \
 			(!memcmp(patch_info->patch_buf + patch_info->patch_len - sizeof(ext_section_sig), ext_section_sig, sizeof(ext_section_sig)))) {
 			HCI_INFO("Signature check success: Merge patch v2");
@@ -976,21 +1046,25 @@ static uint8_t hci_platform_get_patch_info(void)
 			project_id = hci_platform_get_patch_project_id(patch_info->patch_buf + patch_info->patch_len - sizeof(ext_section_sig));
 			if (project_id != HCI_PATCH_PROJECT_ID) {
 				HCI_ERR("Project ID 0x%02x check fail, No available patch!", project_id);
-				return HCI_IGNORE;
+				ret = HCI_IGNORE;
+				goto EXIT;
 			}
 		} else {
 			HCI_ERR("Signature check fail, No available patch!");
-			return HCI_IGNORE;
+			ret = HCI_IGNORE;
+			goto EXIT;
 		}
 	} else {
 		if (!memcmp(patch_info->patch_buf, patch_sig_v1, sizeof(patch_sig_v1))) {
 			HCI_WARN("Signature check success: Merge patch v1 not support");
-			return HCI_IGNORE;
+			ret = HCI_IGNORE;
+			goto EXIT;
 		} else if (!memcmp(patch_info->patch_buf, patch_sig_v2, sizeof(patch_sig_v2))) {
 			HCI_INFO("Signature check success: Merge patch v2");
 		} else {
 			HCI_ERR("Signature check fail, No available patch!");
-			return HCI_IGNORE;
+			ret = HCI_IGNORE;
+			goto EXIT;
 		}
 	}
 
@@ -1001,7 +1075,8 @@ static uint8_t hci_platform_get_patch_info(void)
 	fw_len = hci_platform_parse_patch(patch_info->patch_buf + sizeof(patch_sig_v2) + sizeof(version_date) + sizeof(version_time), &patch_info->head_node);
 	if (fw_len == 0) {
 		HCI_ERR("Available patch not found!");
-		return HCI_IGNORE;
+		ret = HCI_IGNORE;
+		goto EXIT;
 	}
 	HCI_INFO("FW Length: %d", (int)fw_len);
 
@@ -1017,7 +1092,15 @@ static uint8_t hci_platform_get_patch_info(void)
 		patch_info->last_pkt = HCI_PATCH_FRAG_SIZE;
 	}
 
-	return HCI_SUCCESS;
+EXIT:
+#ifdef CONFIG_SECOND_FLASH_PARTITION
+    /* After patch download complete, free the fw patch space */
+	if (rtlbt_fw_ptr) {
+		osif_mem_free(rtlbt_fw_ptr);
+	}
+#endif
+
+	return ret;
 }
 
 uint8_t hci_platform_get_patch_cmd_len(uint8_t *cmd_len)
